feat: Add user ID to createNewTaskIntoDB function
This commit modifies the createNewTaskIntoDB function in the task.service.ts file to include the user ID when creating a new task. The user ID is obtained from the request object and used to assign the task to the corresponding user. This change ensures that tasks are properly associated with the correct user.